+++ changes.md
@@ -32,6 +32,7 @@
 27. Use canvas Theta/Phi angles to set initial camera position in 3D plots 
 28. Create statbox when drawing TGraph with TF1 in list of functions
 29. Implement drawing of TSpline3 and TSpline5 classes 
+30. Handle #color[N] and #splitline{}{} tags in the TLatex
 
 
 ## Changes in 5.2.0